Separate sources1.cs to own files
