+++ DocumentFormat.OpenXml/Packaging/MailMergeRecipientDataPart.cs
@@ -1,0 +1,119 @@
+﻿// Copyright (c) Microsoft Open Technologies, Inc.  All rights reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using DocumentFormat.OpenXml.Office.Word;
+
+namespace DocumentFormat.OpenXml.Packaging
+{
+    /// <summary>
+    /// Defines MailMergeRecipientDataPart.
+    /// </summary>
+    public partial class MailMergeRecipientDataPart : OpenXmlPart
+    {
+        [System.Diagnostics.DebuggerBrowsable(System.Diagnostics.DebuggerBrowsableState.Never)]
+        private DocumentFormat.OpenXml.OpenXmlPartRootElement _rootEle;
+
+        /// <summary>
+        /// Only for OpenXmlPart derived classes.
+        /// </summary>
+        internal override OpenXmlPartRootElement _rootElement
+        {
+            get
+            {
+                return _rootEle;
+            }
+
+            set
+            {
+                _rootEle = value as DocumentFormat.OpenXml.OpenXmlPartRootElement;
+            }
+        }
+
+        /// <summary>
+        /// Gets the root element of this part. The DOM tree will be loaded on demand.
+        /// </summary>
+        internal override OpenXmlPartRootElement PartRootElement
+        {
+            get
+            {
+                if (this.Recipients != null)
+                {
+                    return this.Recipients;
+                }
+                else
+                {
+                    return this.MailMergeRecipients;
+                }
+            }
+        }
+
+        /// <summary>
+        /// Gets/Sets the part's root element when the part's content type is MailMergeRecipientDataPartType.OpenXmlMailMergeRecipientData. 
+        /// Setting this property will throw InvalidOperationException if the MailMergeRecipients property is not null.
+        /// </summary>
+        public DocumentFormat.OpenXml.Wordprocessing.Recipients Recipients
+        {
+            get
+            {
+                TryLoadRootElement();
+                return _rootEle as DocumentFormat.OpenXml.Wordprocessing.Recipients;
+            }
+
+            set
+            {
+                if (value == null)
+                {
+                    throw new ArgumentNullException(nameof(value));
+                }
+                if (this.MailMergeRecipients != null)
+                {
+                    throw new InvalidOperationException(string.Format(System.Globalization.CultureInfo.CurrentCulture, ExceptionMessages.PropertyMutualExclusive, "Recipients", "MailMergeRecipients"));
+                }
+                SetDomTree(value);
+            }
+        }
+
+        /// <summary>
+        /// Gets/Sets the part's root element when the part's content type is MailMergeRecipientDataPartType.MsWordMailMergeRecipientData. 
+        /// Setting this property will throw InvalidOperationException if the Recipients property is not null.
+        /// </summary>
+        public DocumentFormat.OpenXml.Office.Word.MailMergeRecipients MailMergeRecipients
+        {
+            get
+            {
+                TryLoadRootElement();
+                return _rootEle as MailMergeRecipients;
+            }
+
+            set
+            {
+                if (value == null)
+                {
+                    throw new ArgumentNullException(nameof(value));
+                }
+                if (this.Recipients != null)
+                {
+                    throw new InvalidOperationException(string.Format(System.Globalization.CultureInfo.CurrentCulture, ExceptionMessages.PropertyMutualExclusive, "MailMergeRecipients", "Recipients"));
+                }
+                SetDomTree(value);
+            }
+        }
+
+        private void TryLoadRootElement()
+        {
+            if (_rootEle == null)
+            {
+                try
+                {
+                    LoadDomTree<DocumentFormat.OpenXml.Wordprocessing.Recipients>();
+                }
+                catch (System.IO.InvalidDataException)
+                {
+                }
+                if (_rootEle == null)
+                {
+                    LoadDomTree<DocumentFormat.OpenXml.Office.Word.MailMergeRecipients>();
+                }
+            }
+        }
+    }
+}
